requirements.txt
# Base
sqlalchemy==1.4.46
alembic==1.13.1
psycopg2-binary==2.9.5

# CLI et interface
typer[all]==0.9.0
rich==13.7.0

# Validation et sérialisation
pydantic==1.10.13

# Sécurité
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Configuration
python-dotenv==1.0.0

# Journalisation
loguru==0.7.2
sentry-sdk==1.38.0

# Tests
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Développement
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0